Rename navigation map folder
