+++ settings/lib/ui/navigation/map_destination_item.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:settings/model/map_destination_model.dart';
-import 'package:settings/ui/map/cubit/map_destination_item_cubit.dart';
-import 'package:settings/ui/map/cubit/map_destination_item_state.dart';
-import 'package:settings/ui/map/map_destination_selection_dialog.dart';
+import 'package:settings/ui/navigation/cubit/map_destination_item_cubit.dart';
+import 'package:settings/ui/navigation/cubit/map_destination_item_state.dart';
+import 'package:settings/ui/navigation/map_destination_selection_dialog.dart';
 
 class MapDestinationItem extends StatelessWidget {
   const MapDestinationItem({Key? key}) : super(key: key);
@@ -28,7 +28,7 @@ class MapDestinationItem extends StatelessWidget {
                   },
                   minLeadingWidth: 10,
                   leading: const Icon(Icons.navigation),
-                  title: const Text("Navigations App"),
+                  title: const Text("Navigation starten mit.."),
                   subtitle: Text(state.valueLabel ?? ""));
             }));
   }
