+++ settings/lib/ui/settings/settings_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:settings/ui/developer/developer_card.dart';
-import 'package:settings/ui/map/map_destination_item.dart';
+import 'package:settings/ui/navigation/map_destination_item.dart';
 import 'package:settings/ui/settings/settings_card.dart';
 import 'package:settings/ui/theme/theme_item.dart';
 import 'package:settings/ui/version/version_item.dart';
